+++ src/date/toUnix.js
@@ -1,0 +1,9 @@
+'use strict';
+
+import isDate from './is';
+
+export default function toUnix (val) {
+    if (!isDate(val)) throw new Error('toUnix requires a date object');
+
+    return Math.floor(val.valueOf()/1000);
+}